Actually think I have fixed it this time. Found that the context for the field was being updated in another place.
So I added the key 'bista_show_address' to the two fields,
partner_invoice_id and partner_shipping_id. Then in the name_get
method, check for this in the context dictionary and set the name
appropriately.

Looking at other places in Odoo, did not find that the customer
display had changed.